+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to this project will be documented in this file.
 
+## v3 (2020-09-13)
+
+- Added Global HTML fields
+
 ## v2 (2020-03-14)
 
 - Fixed compatibility with Mobirise v5